+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ## History
 
+### 0.2.0 (2022 Sep 22)
+
+ - new: `TimeWrapper` interface for time unit classes
+ - new: `Timestamp` wrapper around timestamps
+ - new: `abs` absolute value function added to both classes
+ - update: `Timespan`: get durations with specified units only
+ - update: `Timespan`: add month unit
+ - update: `Timespan`: **symbolType** renamed to **symbolFormat**
+ - update: `Timespan`: moved moment functions to timestamp
+ - update: `Timespan`: handles historic/negative values
+ - update: `Timespan`: format now returns duration
+
 ### 0.1.0 (2022 Aug 13)
 
  - initial